changes according to Frends Task development guide
