+++ Frends.OPCUA.WriteTags/Frends.OPCUA.WriteTags/Frends.OPCUA.WriteTags.Tests/Integrations/OpcUaClientIntegrationTests.cs
@@ -1,10 +1,13 @@
-﻿namespace Frends.OPCUA.WriteTags.Tests.Integrations;
+﻿using Frends.OPCUA.WriteTags.Definitions;
+using Frends.OPCUA.WriteTags.Services;
+
+namespace Frends.OPCUA.WriteTags.Tests.Integrations;
 
 public class OpcUaClientIntegrationTests
 {
     private IOpcUaClientFactory _factory = new OpcUaClientFactory();
 
-    private WriteValue _writeInt = new WriteValue
+    private WriteTagValue _writeTagInt = new WriteTagValue
     {
         NodeId = "ns=2;s=Scalar_Static_Int32",
         Value = 100
@@ -27,7 +30,7 @@ public class OpcUaClientIntegrationTests
 
         await Assert.ThrowsAsync<OpcSessionException>(async () =>
         {
-            await sut.WriteDataAsync(new[] { _writeInt });
+            await sut.WriteDataAsync(new[] { _writeTagInt });
         });
     }
 
@@ -46,13 +49,13 @@ public class OpcUaClientIntegrationTests
         };
         var sut = _factory.CreateClient(input, options);
         
-        var response = await sut.WriteDataAsync(new[] { _writeInt });
+        var response = await sut.WriteDataAsync(new[] { _writeTagInt });
 
         Assert.NotNull(response);
         Assert.True(response.IsAllSuccess);
         Assert.NotEmpty(response.Results);
-        Assert.Equal(_writeInt.NodeId, response.Results[0].NodeId);
-        Assert.Equal(_writeInt.Value, response.Results[0].Value);
+        Assert.Equal(_writeTagInt.NodeId, response.Results[0].NodeId);
+        Assert.Equal(_writeTagInt.Value, response.Results[0].Value);
     }
     
     [Fact]
@@ -72,13 +75,13 @@ public class OpcUaClientIntegrationTests
         };
         var sut = _factory.CreateClient(input, options);
         
-        var response = await sut.WriteDataAsync(new[] { _writeInt });
+        var response = await sut.WriteDataAsync(new[] { _writeTagInt });
 
         Assert.NotNull(response);
         Assert.True(response.IsAllSuccess);
         Assert.NotEmpty(response.Results);
-        Assert.Equal(_writeInt.NodeId, response.Results[0].NodeId);
-        Assert.Equal(_writeInt.Value, response.Results[0].Value);
+        Assert.Equal(_writeTagInt.NodeId, response.Results[0].NodeId);
+        Assert.Equal(_writeTagInt.Value, response.Results[0].Value);
     }
     
     [Fact]
@@ -98,7 +101,7 @@ public class OpcUaClientIntegrationTests
 
         await Assert.ThrowsAnyAsync<Exception>(async () =>
         {
-            await sut.WriteDataAsync(new[] { new WriteValue { NodeId = "invalid_node_id", Value = "42" } });
+            await sut.WriteDataAsync(new[] { new WriteTagValue { NodeId = "invalid_node_id", Value = "42" } });
         });
     }
 }
